refactor preview code
